Work around ~/.cache/ permission issues.
root-owned files get created in NB_USER's ~/.cache/ directory. We carefully try to step around the issue. I've see the directory get created as ~/.cache/rosetta, which makes me think it is an artifact of the local docker builder rather than the contents of the Dockerfile or any install script.